ai-engine/requirements.txt
# ─── Web Framework ───────────────────────────────────────────
fastapi>=0.110.0
uvicorn[standard]>=0.27.1
python-multipart>=0.0.9

# ─── Document Parsing ────────────────────────────────────────
pdfplumber>=0.11.0
python-docx>=1.1.0
chardet>=5.2.0

# ─── Local LLM Inference (Ollama client) ─────────────────────
httpx>=0.27.0
ollama>=0.1.8

# ─── Local Embeddings (fully offline, no API key needed) ─────
sentence-transformers>=2.7.0
# NOTE: Install torch BEFORE this file via:
#   py -3.13 -m pip install torch --index-url https://download.pytorch.org/whl/cpu

# ─── NLP utilities ───────────────────────────────────────────
spacy>=3.7.4
# Download model separately: py -3.13 -m spacy download en_core_web_sm

# ─── Data / Validation ───────────────────────────────────────
pydantic>=2.6.3
python-dotenv>=1.0.1
numpy>=1.26.4
scikit-learn>=1.4.1

# ─── Async / Utilities ───────────────────────────────────────
aiofiles>=23.2.1
python-jose>=3.3.0
passlib>=1.7.4
loguru>=0.7.2

# ─── Fine-tuning (optional, install separately on GPU machine) ─
# unsloth[colab-new]
# trl
# peft
# bitsandbytes
# datasets
# transformers>=4.38.0
